moved dockLigands to MoltimateContainer
 - moved dockLigands because useForm needs to be called from inside a React component (because it uses state hooks)
